Extract common schedule-plan helpers to core/models/common/utils.py
Move the model-agnostic schedule-plan pieces — weak_method,
should_free_input, TransformerLayerState, TransformerLayerNode,
_BackwardDWWrapper, PreProcessNode, PostProcessNode — out of
core/models/gpt/fine_grained_callables.py and into a new
core/models/common/utils.py. The common chunk-schedule-plan module and
the hybrid path now import from common/utils.py instead of crossing
into gpt/. The GPT-specific TransformerLayer / MTPLayer callable
builders stay in gpt/fine_grained_callables.py and re-export the moved
names so existing imports keep working.

Pre/PostProcessNode now use a generic ``model`` attribute (was
``gpt_model``); they call ``model._preprocess`` / ``model._postprocess``
which works for any model that exposes those methods. With the rename
HybridModel is just another consumer of the same nodes — the dedicated
HybridPreProcessNode / HybridPostProcessNode subclasses are no longer
needed and are deleted, along with the
PRE_PROCESS_NODE_CLASS / POST_PROCESS_NODE_CLASS overrides on
HybridStackModelChunkSchedulePlan.

transformer/module.py also imports _BackwardDWWrapper from common/utils.py
now; the previous import path crossed from megatron/core/transformer
into megatron/core/models/gpt/, which is the cross-cut the reviewer
flagged.
